Move PDODriverTrait methods into Driver class itself.
